% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_unequal_vectors.R
\name{add_unequal_vectors}
\alias{add_unequal_vectors}
\title{Add vectors of unequal length}
\usage{
add_unequal_vectors(x, y)
}
\arguments{
\item{x, y}{Paired numeric vectors of same or different length.}
}
\value{
A vector of length one containing pre-spawn survival.
}
\description{
Function used to add aligned vectors of unequal length without
recycling elements of the shorter vector. Used to combine male and
female cohorts when \code{max_age} is not equal in
the \code{\link{max_ages}} dataset.
}
\examples{
x <- c(1, 2)
y <- c(1, 2, 3)
add_unequal_vectors(x, y)

}
\references{
https://stackoverflow.com/questions/2307443/how-to-add-two-vectors-without-repeating-in-r
}
