% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crecco_1983.R
\docType{data}
\name{crecco_1983}
\alias{crecco_1983}
\title{Daily mortality of juvenile American shad}
\format{
A data frame with 24 observations of 6 variables:
\describe{
    \item{\code{Year}}{Year of study}
    \item{\code{Age}}{Fish age (d)}
    \item{\code{Md}}{Daily mortality}
    \item{\code{Sd}}{Daily survival}
    \item{\code{Si}}{Interval survival}
    \item{\code{Sc}}{Cumulative survival}
}

#' @references Crecco, V., T. Savoy, and L. Gunn. 1983. Daily mortality rates 
of larval and juvenile American shad (*Alosa sapidissima*) in the Connecticut 
River with changes in year-class strength. Canadian Journal of 
Fisheries and Aquatic Sciences 40:1719–1728.
}
\usage{
crecco_1983
}
\description{
A built-in dataset containing estimated daily 
mortality rates for larval and juvenile American shad in the
Connecticut River, USA 1979-1982 (Crecco et al. 1983).
}
\keyword{datasets}
