% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_tl_conversions.R
\docType{data}
\name{fl_tl_conversions}
\alias{fl_tl_conversions}
\title{Fork length - total length conversions for river herring}
\format{
A data frame with 2 observations of 5 variables:
\describe{

\code{species} Species of river herring ("ALE" or "BBH")

\code{alpha} Intercept

\code{alpha.se} Standard error of the intercept

\code{beta} Slope

\code{beta.se} Standard error of the slope

}
}
\source{
Atlantic States Marine Fisheries Commission
}
\usage{
fl_tl_conversions
}
\description{
A dataset containing species-specific fork length - total
length conversions from ASMFC (2024). The data are regression
coefficients and standard errors for each species.
}
\references{
Atlantic States Marine Fisheries Commission. 2024. River herring
benchmark stock assessment and peer-review report. ASMFC, Arlington, VA. 
URL: https://asmfc.org/uploads/file/66f59e40RiverHerringAssessment_PeerReviewReport_2024.pdf
}
\keyword{datasets}
