% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dams.R
\name{get_dams}
\alias{get_dams}
\title{Get dams for specified river from American shad habitat data}
\usage{
get_dams(river)
}
\arguments{
\item{river}{Character string specifying river name}
}
\value{
a data.frame with 4 variables containing dam name,
latitude and longitude, and dam order in the watershed
}
\description{
Function used to get dams for rivers listed
in \code{\link{get_rivers}} from the built-in 
data set(s) for American shad. Not implemented for river herring
}
\examples{
get_dams("Penobscot")

}
