% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region.R
\name{get_region}
\alias{get_region}
\title{Get region for specified river by species}
\usage{
get_region(
  river,
  species = c("ALE", "AMS", "BBH", "EEL"),
  custom_habitat = NULL
)
}
\arguments{
\item{river}{Character string specifying river name.}

\item{species}{Character string specifying species.
Choices include American shad (\code{"AMS"}), alewife (\code{"ALE"}), and
blueback herring (\code{"BBH"}, and American eel (\code{"EEL"})).}

\item{custom_habitat}{A dataframe containing columns corresponding to the
those in the output from \code{\link{custom_habitat_template}}. The default,
\code{NULL} uses the default habitat data set for a given combination of
\code{species} and \code{river}.}
}
\description{
Function used to get region for rivers listed
in output of \code{\link{get_rivers}} from the built-in
habitat data sets.
}
\examples{
get_region(river = "Hudson", species = "AMS")

}
