% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rivers.R
\name{get_rivers}
\alias{get_rivers}
\title{Rivers included in habitat data by species}
\usage{
get_rivers(species = c("ALE", "AMS", "BBH", "EEL"))
}
\arguments{
\item{species}{Species for which rivers are returned
Choices include American shad (\code{"AMS"}), alewife (\code{"ALE"}), 
blueback herring (\code{"BBH"}, and American eel (\code{"EEL"})).}
}
\value{
A character vector of rivers that are included in
built-in habitat datasets for selected \code{species}.
}
\description{
Get names of rivers included in package for each species.
}
\details{
The primary uses of this function are (1) so the user can see
rivers that can be used for simulation, and (2) so the output can be passed
to a random sampler for use in stochastic simulation via
parallel processing.
}
\examples{
# Example usage
\dontrun{
  
# Use get_rivers() to return names of rivers included in package
get_rivers(anadrofish::habitat)

# Sample a random river from those available
get_rivers(anadrofish::habitat)[sample(
  length(get_rivers(anadrofish::habitat)), 1)]  
  
}
}
