% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_ale.R
\docType{data}
\name{habitat_ale}
\alias{habitat_ale}
\title{Alewife habitat}
\format{
A data frame with 246,259 observations of 7 variables:
\describe{

\code{REACHCODE}{ code used to identify individual habitat units}

\code{Hab_sqkm}{ surface area of habitat in square kilometers}

\code{Latitude}{ latitude at downstream terminus of habitat unit}

\code{State}{ governmental unit at downstream terminus of habitat unit}

\code{River_huc}{ name of river to which habitat unit belongs, derived from hydrologic unit codes}

\code{POP}{ genetic reporting group geographic region within which habitat unit falls}

\code{DamOrder}{ order of dam at downstream terminus of habitat unit. Cumulatively assigned such that all habitat units upstream of a given dam all have dam_order >= 1}

}
}
\source{
Shawn Snyder
}
\usage{
habitat_ale
}
\description{
A dataset containing the surface area of Alewife 
habitat and features for individual flow line segments or waterbodies (lakes)
comprising habitat units in Atlantic coast drainages.
}
\examples{
head(habitat_ale)


}
\keyword{datasets}
