% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower95.R
\name{lower95}
\alias{lower95}
\title{Calculate lower 95 \% confidence limit}
\usage{
lower95(x)
}
\arguments{
\item{x}{A numeric vector or matrix.}
}
\value{
Numeric vector of length 1 containing the 2.5th percentile of the sampling distribution for \code{x}.
}
\description{
Function with a short name to calculate 2.5th percentile
of a distribution. Calls \code{stats::quantile(x)}, 
but is shorter and is easier to use in calls to \code{apply} family
of functions.
}
\examples{
# Example usage
\dontrun{
  
 x <- rnorm(1e4, 0, 1)
 
 lower95(x)
  
}
}
