% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lw_pars_rh.R
\docType{data}
\name{lw_pars_rh}
\alias{lw_pars_rh}
\title{Length-weight regression coefficients for river herring}
\format{
A data frame with 30 observations of 7 variables:
\describe{

\code{Species}{ species of river herring ("ALE" or "BBH")}

\code{sex}{ Fish sex}

\code{alpha}{ Intercept}

\code{alpha.se}{ Standard error for the intercept}

\code{beta}{ Slope}

\code{beta.se}{ Standard error for the slope}

}
}
\source{
Atlantic States Marine Fisheries Commission
}
\usage{
lw_pars_rh
}
\description{
A dataset containing sex-specific regression
coefficients of log10length-log10weight relationships
for each genetic reporting group of river herring.
}
\keyword{datasets}
