% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_biomass_in.R
\name{make_biomass_in}
\alias{make_biomass_in}
\title{Make biomass in}
\usage{
make_biomass_in(spawners, spawners_down, mass)
}
\arguments{
\item{spawners}{The number of spawner recruits.}

\item{spawners_down}{The number of spawners that successfully spawned in
freshwater and returned to the ocean.}
}
\value{
Biomass in grams of unsuccessful spawners. Numeric vector 
of \code{length(spawners)}.
}
\description{
Function used to calculate biomass transported into freshwater.
}
\examples{
# Example usage
\dontrun{
    
    # Generated values 
    spawners <- c(0, 0, 0, 1634, 14506, 11674, 6392, 3521, 1948)
    spawners_down <- c(0, 0, 0, 1464, 13064, 10542, 5782, 3189, 1765)
    mass <- c(64, 144, 206, 247, 271, 284, 292, 296, 298)   
    
    # Calculate expected number of age-0 fish  
    biomass_in <- make_biomass_in(spawners = spawners, spawners_down = spawners_down, mass = mass)
    
}
}
