% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_biomass_out.R
\name{make_biomass_out}
\alias{make_biomass_out}
\title{Make biomass out}
\usage{
make_biomass_out(age0_down)
}
\arguments{
\item{age0_down}{The number of successful juvenile outmigrants.}
}
\value{
Biomass in grams of successful juvenile outmigrants. Numeric vector 
is a single value (vector of length 1).
}
\description{
Function used to calculate biomass transported out of freshwater.
}
\examples{
# Example usage
\dontrun{
    
    # Generated values 
    age0_down <- 1000000  
    
    # Calculate expected number of age-0 fish  
    biomass_out <- make_biomass_in(age0_down = age0_down)
    
}
}
