% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eggs.R
\name{make_eggs}
\alias{make_eggs}
\title{Simulate number of eggs per fish}
\usage{
make_eggs(
  river,
  species = c("AMS", "ALE", "BBH", "EEL"),
  custom_habitat = NULL
)
}
\arguments{
\item{river}{The river for which eggs will be simulated.}

\item{species}{Species for which population dynamics will be simulated.
Choices include American shad (\code{"AMS"}), alewife (\code{"ALE"}), 
blueback herring (\code{"BBH"}), and American eel (\code{"EEL"}).}

\item{custom_habitat}{A dataframe containing columns corresponding to the
those in the output from \code{\link{custom_habitat_template}}. The default,
\code{NULL} uses the default habitat data set for a given combination of
\code{species} and \code{river}.}
}
\value{
A vector containing age-specific potential annual
fecundity with \code{length = max_age}.
}
\description{
Function used to simulate number of eggs
produced per female (potential annual fecundity, PAF),
for rivers included in \code{\link{get_rivers}}.
}
\section{Details}{

The default method for American shad uses predictive equations from Olney and
McBride (2003) to simulate batch fecundity from weight-fecundity relationships
using coefficients in \code{\link{olney_mcbride}} for the life-history region
corresponding to the specified river. Number of batches for this
approach is currently drawn from a random normal distribution
with a mean of 6.7 and a standard deviation of 2.1 (McBride et al. 2016).

The default method for river herring calls \code{\link{make_eggs_rh}} with
simulation methods and references described therein.
}

\examples{
make_eggs(river = "Susquehanna", species = "AMS")

}
\references{
Olney, J. E. and R. S. McBride. 2003. Intraspecific
variation in batch fecundity of American shad (Alosa sapidissima):
revisiting the paradigm of reciprocal latitudinal trends
in reproductive traits. American Fisheries Society
Symposium 35:185-192.

McBride, R. S., R. Ferreri, E. K. Towle, J. M. Boucher, and
G. Basilone, G. 2016. Yolked oocyte dynamics support agreement
between determinate- and indeterminate-method estimates of annual
fecundity for a northeastern United States population of
American shad. PLoS ONE 11(10): 10.1371/journal.pone.0164203
}
