% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eggs_eel.R
\name{make_eggs_eel}
\alias{make_eggs_eel}
\title{Simulate number of eggs per fish for American eel}
\usage{
make_eggs_eel(river, species = "EEL", length, custom_habitat = NULL)
}
\arguments{
\item{species}{Species for which population dynamics will be simulated.
Choice is American eel (\code{"EEL"}).}

\item{custom_habitat}{A dataframe containing columns corresponding to the
those in the output from \code{\link{custom_habitat_template}}. The default,
\code{NULL} uses the default habitat data set for a given combination of
\code{species} and \code{river}.}
}
\value{
A vector containing age-specific potential annual
fecundity with \code{length = max_age}.
}
\description{
Function used to simulate number of eggs
produced per female (total fecundity) for American eel.
}
\section{Details}{

The default method for American eel uses a stochastic, MCMC sampling 
approach to draw correlated sets of von Bertalanffy growth
parameters in \code{vbgf_eel}, length-weight regression parameters in 
\code{link{lw_pars_eel}}, and length-fecundity relationships from ASFMC 
(2011), Barbin & McCleave (1997), Tremblay (2009), and Wenner & Musick (1974)
to simulate the number of eggs per female within regional reporting groups
(\code{region}).
}

\examples{
make_eggs_rh(river = "Susquehanna", species = "BBH")

}
\references{
Atlantic States Marine Fisheries Commission. 2024. American eel
benchmark stock assessment and peer-review report. ASMFC, Arlington, VA.
URL:
}
