% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eggs_rh.R
\name{make_eggs_rh}
\alias{make_eggs_rh}
\title{Simulate number of eggs per fish for river herring}
\usage{
make_eggs_rh(river, species = c("ALE", "BBH"), custom_habitat = NULL)
}
\arguments{
\item{river}{The river for which eggs will be simulated.}

\item{species}{Species for which population dynamics will be simulated.
Choices include alewife (\code{"ALE"}), and blueback herring (\code{"BBH"}).}

\item{custom_habitat}{A dataframe containing columns corresponding to the
those in the output from custom_habitat_template(). NEED TO ADD LINK.}
}
\value{
A vector containing age-specific potential annual 
fecundity with \code{length = max_age}.
}
\description{
Function used to simulate number of eggs 
produced per female (total fecundity) for river herring in
rivers included in \code{\link{get_rivers}}.
}
\section{Details}{
 NEED TO ADD DETAILS.
}

\examples{
make_eggs_rh(river = "Upper Susquehanna", species = "BBH")

}
\references{
Sullivan et al. (2019), Jessop (1993)
}
