% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fec.R
\name{make_fec}
\alias{make_fec}
\title{Make fecundity}
\usage{
make_fec(eggs, sr, s_hatch)
}
\arguments{
\item{eggs}{The number of eggs an individual fish
can produce}

\item{sr}{Sex ratio}

\item{s_hatch}{Hatching success (hatch to larval survival)}
}
\value{
Number of age-0 recruits produced. May be a single 
value (vector of length 1) or a vector of length = maximum age
depending on whether age-specific values are passed to \code{eggs}
}
\description{
Function used to make fecundity values
for age-structured spawning population of fish
}
\examples{
# Example usage
\dontrun{
  
# Using values from Bailey and Zydlewski (2013)  
  eggs <- c(0, 0, 0, 20654, 34674, 58210, 79433, 88480, 97724)
  sr <- 0.50
  s_hatch = 0.10   

# Calculate expected number of age-0 fish  
  fec <- make_fec(eggs = eggs, sr = sr, s_hatch = s_hatch)
  
}
}
\references{
Bailey, M. M., and J. D. Zydlewski. 2013. To stock or not
to stock? Assessing the restoration potential of a remnant American shad
spawning run with hatchery supplementation. North American Journal of
Fisheries Management 33:459-467. doi: 10.1080/02755947.2013.763874.
}
