% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_iteroparity.R
\name{make_iteroparity}
\alias{make_iteroparity}
\title{Estimate rate of iteroparity for American shad from
latitude based on Leggett and Cascardden (1978) and 
Bailey and Zydlewski (2014)}
\usage{
make_iteroparity(latitude)
}
\arguments{
\item{habitat}{A built-in \code{data.frame} containing habitat 
estimates (km2) for the scenario and region selected.}
}
\description{
Function used to extract river-specific latitude
from built-in data sets.
}
