% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mass.R
\name{make_mass}
\alias{make_mass}
\title{Mass by age in American eel.}
\usage{
make_mass(river, species = "EEL", length, custom_habitat = NULL)
}
\arguments{
\item{species}{Species for which population dynamics will be simulated.
Choice is American eel (\code{"EEL"}).}

\item{custom_habitat}{A dataframe containing columns corresponding to the
those in the output from \code{\link{custom_habitat_template}}. The default,
\code{NULL} uses the default habitat data set for a given combination of
\code{species} and \code{river}.}
}
\value{
A vector containing age-specific mass (g) \code{length = max_age}.
}
\description{
Function used to simulate Mass by age for American eel.
}
\section{Details}{

The default method for American eel uses a stochastic approach to draw 
allometric growth parameters from \code{lw_pars_eel}.
}

\references{
Atlantic States Marine Fisheries Commission. 2024. American eel
benchmark stock assessment and peer-review report. ASMFC, Arlington, VA.
URL: Table 2.
}
