% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_output.R
\name{make_output}
\alias{make_output}
\title{Make output vectors}
\usage{
make_output()
}
\arguments{
\item{params}{Character vector of parameters}
}
\description{
Create output vectors for default
arguments in \code{\link{sim_pop}}.

Not intended to be called directly.
}
