% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pfas_in.R
\name{make_pfas_in}
\alias{make_pfas_in}
\title{Make PFAS in}
\usage{
make_pfas_in(spawners, spawners_down, mass)
}
\arguments{
\item{spawners}{The number of spawner recruits.}

\item{spawners_down}{The number of spawners that successfully spawned in
freshwater and returned to the ocean.}

\item{mass}{Mass at age vector}
}
\value{
Total PFAS, here PFOS, in nanograms offloaded by spawner recruit carcasses into freshwater.
Essentially those that die from post-spawning, dam death, and from 2 months of natural mortality.
Numeric value of \code{length(spawners)}. Simulated concentrations per individual from Melnyk et al. 2024. Migrating BBH from Penobscot River in 2021.
}
\description{
Function used to simulate total PFAS transported into freshwater.
}
\examples{
# Example usage
\dontrun{
    
    # Generated values 
    spawners <- c(0, 0, 0, 1634, 14506, 11674, 6392, 3521, 1948)
    spawners_down <- c(0, 0, 0, 1464, 13064, 10542, 5782, 3189, 1765)
    mass <- c(64, 144, 206, 247, 271, 284, 292, 296, 298)   
    
    # Calculate expected number of age-0 fish  
    pfas_in <- make_pfas_in(spawners = spawners, spawners_down = spawners_down, mass = mass)
    
}
}
