% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pfas_out.R
\name{make_pfas_out}
\alias{make_pfas_out}
\title{Make PFAS out}
\usage{
make_pfas_out(age0_down)
}
\arguments{
\item{age0_down}{Number of outmigrating juveniles}
}
\value{
Total PFAS, here PFOS, in nanograms transported by outmigrating juveniles from freshwater to ocean.
Numeric vector is a single value (vector of length 1).
Simulated concentrations per individual from Xia et al. 2024. Resident adult alewife in Lake Michigan from 1994.
}
\description{
Function used to simulate total PFAS transported out of freshwater.
}
\examples{
# Example usage
\dontrun{
    
    # Generated values 
    age0_down <- 1000000  
    
    # Calculate expected number of age-0 fish  
    pfas_out <- make_pfas_out(age0_down = age0_down)
    
}
}
