% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_postspawn.R
\name{make_postspawn}
\alias{make_postspawn}
\title{Post-spawning survival}
\usage{
make_postspawn(
  river = river,
  species = c("AMS", "ALE", "BBH", "EEL"),
  iteroparity = NULL,
  nM = NULL,
  custom_habitat = NULL
)
}
\arguments{
\item{river}{River for which post-spawn survival rate should be
returned. Required argument with no default value. Available rivers
can be seen using \code{\link{get_rivers}}.}

\item{species}{Species for which population dynamics will be simulated.
Choices include American shad (\code{"AMS"}), alewife (\code{"ALE"}),
blueback herring (\code{"BBH"}), and American eel (\code{"EEL"}).}

\item{iteroparity}{Optional argument for rate of iteroparity. Values from
\code{\link{make_iteroparity}} can be passed directly to this function, or
a numeric vector of \code{length = 1}.}

\item{nM}{Instantaneous annual mortality. Values for
natural mortality for life-history regions can be
from \code{\link{mortality}} for \code{"AMS"}, \code{\link{mortality_rh}} for
\code{"ALE"} and \code{"BBH"}, or a numeric vector of \code{length = 1}.}

\item{custom_habitat}{A dataframe containing columns corresponding to the
those in the output from \code{\link{custom_habitat_template}}. The default,
\code{NULL} uses the default habitat data set for a given combination of
\code{species} and \code{river}.}
}
\description{
Function used to estimate post-spawn survival
from proportion of repeat spawners by latitude (Leggett and
Cascardden 1978, Bailey and Zydlewski 2013) and natural mortality
by life-history region ("AMS") or by iteroparity and natural mortality
("ALE" and "BBH").
}
\examples{
make_postspawn(river = "Susquehanna", species = "AMS")

}
\references{
Bailey, M.M., and J. D. Zydlewski. 2013. To stock or not
to stock? Assessing the restoration potential of a remnant
American shad spawning run with hatchery supplementation. North
American Journal of Fisheries Management 33:459–467.

Leggett, W., and J. E. Cascardden. 1978. Latitudinal Variation in
Reproductive Characteristics of American Shad (Alosa sapidissima):
Evidence for Population Specific Life History Strategies in Fish.
Journal of the Fisheries Research Board of Canada 35:1469-1478.

Atlantic States Marine Fisheries Commission. 2024. River herring
benchmark stock assessment and peer-review report. ASMFC, Arlington, VA.
URL: https://asmfc.org/uploads/file/66f59e40RiverHerringAssessment_PeerReviewReport_2024.pdf
}
