% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prespawn.R
\name{make_prespawn}
\alias{make_prespawn}
\title{Make pre-spawning mortality}
\usage{
make_prespawn(nM, s_postspawn)
}
\arguments{
\item{nM}{Instantaneous natural mortality rate. 
A numeric vector of length 1.}

\item{s_postspawn}{Post spawning mortality calculated from the
output of make_iteroparity and nM}
}
\value{
A vector of length one containing pre-spawn survival.
}
\description{
Function used to derive pre-spawning survival from
natural mortality rate (river and/or sex-specific) and post-spawning
survival (sex-specific or not) based on river-specific iteroparity.
}
