% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_recruits.R
\name{make_recruits}
\alias{make_recruits}
\title{Make recruits}
\usage{
make_recruits(eggs, sr)
}
\arguments{
\item{eggs}{The number of eggs an individual fish
can produce (Potential annual fecundity, PAF).}

\item{sr}{Sex ratio of spawning adults.}
}
\value{
Number of juvenile recruits. Numeric vector of
\code{length(eggs}. May be a single value (vector of length 1)
or a vector of \code{length = maximum age} depending on whether
age-specific values are passed to \code{eggs}.
}
\description{
Function used to calculate realized reproductive output 
(to larval stage) of female spawners without density-dependence
in larval production.
}
\examples{
# Example usage
\dontrun{
  
# Using values from Bailey and Zydlewski (2013)  
  eggs <- c(0, 0, 0, 20654, 34674, 58210, 79433, 88480, 97724)
  sr <- 0.50
  s_hatch = 0.10   

# Calculate expected number of age-0 fish  
  fec <- make_recruits(eggs = eggs, sr = sr, s_hatch = s_hatch)
  
}
}
