% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_s_spawn.R
\name{make_s_spawn}
\alias{make_s_spawn}
\title{Make total spawning migration mortality}
\usage{
make_s_spawn(nM, s_postspawn)
}
\arguments{
\item{nM}{Instantaneous natural mortality rate.
A numeric vector containing either an age-invariant natural mortality
rate of length 1 or an age-specific natural mortality rate with
length = max_age.}

\item{s_postspawn}{Post spawning mortality calculated from the
output of \code{\link{make_iteroparity}} and \code{nM}, or alternatively a
numeric vector for each.}
}
\value{
A vector of length one containing pre-spawn survival.
}
\description{
Function used to derive total in-river adult survival during
the spawning migration from annual mortality and post-spawn survival. This function
is not currently implemented in the underlying population model, can be used
to estimate numbers of fish lost in freshwater environments for purposes such
as nutrient modeling.
}
\examples{
make_s_spawn(0.68, 0.92)

}
