% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spawnrecruit_rh.R
\name{make_spawnrecruit_rh}
\alias{make_spawnrecruit_rh}
\title{Make spawnrecruit vectors for river herring}
\usage{
make_spawnrecruit_rh(
  river,
  sex = c("male", "female"),
  species = c("ALE", "BBH")
)
}
\arguments{
\item{sex}{A character indicating \code{"male"}, \code{"female"},
or \code{"Pooled"} sex for fish.}

\item{species}{A character indicating species of river herring, either
\code{"ALE" for alewife or "BBH" for blueback herring}.}

\item{region}{A character indicating genetic stock groupings for 
alewife (one of "MAT", "SNE", or "NNE") or blueback herring (one of
one of "SAT", "MAT", "SNE", "MNE", or "CAN-NNE") in 
in \code{anadrofish::maki_pars}.}
}
\value{
A numeric vector containing a single realization for proportion
of mature fish at each age, from age 1 to maximum age.
}
\description{
Simulate proportion of population that is mature
spawners from sex-specific maturity schedules for river herring
in \code{\link{maki_pars}}.
}
\details{
The primary use of this function is to simulate proportion of
mature spawners at each age in a population of river-herring based on 
region-specific probabilities of maturation at each age estimated
from the Maki (YEAR) method in ASMFC (2024).
}
\examples{
# Example usage
\dontrun{
  
  # Use make_spawnrecruit_rh() to simulate proportion of mature
  # fish at each age
  spawn_recruit_vector <- make_spawnrecruit_rh(
    species = "ALE", region = "NNE", sex = "Male")
  
  plot(spawn_recruit_vector, type = "l")
  
}
}
