% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_upstream.R
\name{make_upstream}
\alias{make_upstream}
\title{Calculate upstream survival given dam passage scenario.}
\usage{
make_upstream(river, species = "EEL", upstream, custom_habitat = NULL)
}
\arguments{
\item{river}{Character string specifying river name.}

\item{species}{Species for which population dynamics will be simulated.
Choices include American eel (\code{"EEL"}).}

\item{upstream}{Numeric indicating proportional upstream passage
through a single dam.}

\item{custom_habitat}{A dataframe containing columns corresponding to the
those in the output from \code{\link{custom_habitat_template}}. The default,
\code{NULL} uses the default habitat data set for a given combination of
\code{species} and \code{river}.}
}
\value{
Numeric vector of length 1 representing weighted catchment-scale
downstream migration mortality for juvenile or adult fish.
}
\description{
Function used to create population-level
survival through dams during seasonal upstream migration of
juveniles.
}
\details{
This function assigns cumulative downstream passage values
to all features in \code{\link{habitat}} corresponding to \code{river}.
It then calculates the proportion of habitat in each habitat segment of a
river, and weights downstream mortality at the catchment-scale
by proportion of habitat. This implicitly assumes that fish are
distributed throughout the river during spawning in proportion to
available habitat.
}
\examples{
# Example usage
\dontrun{
  
# Example 1. ----  
# Calculate juvenile survival during in-migration
# for fixed upstream and downstream dam passage probabilities
s_up <- make_upstream(habitat_data = habitat,
                          river = 'Susquehanna',
                          upstream = 0.80)
}
}
