% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturity.R
\docType{data}
\name{maturity}
\alias{maturity}
\title{Sex-specific maturity schedules for American shad}
\format{A data frame with 65 observations of 6 variables:
\describe{

\code{region}{ Life-history region}

\code{sex}{ Fish sex (gender)}

\code{age1, age2, ..., age13}{ Probability of maturing at age}
}}
\source{
Atlantic States Marine Fisheries Commission
}
\usage{
maturity
}
\description{
A dataset containing sex-specific probabilities 
of maturation at age for each life-history region 
(NI = Northern iteroparous, SI = Southern iteroparous, 
SP = Semelparous)
}
\keyword{datasets}
