% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_ages_eel.R
\docType{data}
\name{max_ages_eel}
\alias{max_ages_eel}
\title{Sex-specific maximum ages of American eel}
\format{
A data frame with 12 observations of 3 variables:
\describe{
    \item{\code{region}}{Life-history region}
    \item{\code{sex}}{Fish sex}
    \item{\code{maxage}}{Maximum observed age}
}
}
\source{
Atlantic States Marine Fisheries Commission
}
\usage{
max_ages_eel
}
\description{
A dataset containing sex-specific maximum
ages for each life-history region (GM = Gulf of Maine, SNE = Southern
New England, HUD = Hudson, DMA = Delaware/Maryland, CHE = Chesapeake, SAT = 
Southern Atlantic
}
\references{
ddkdkd
}
\keyword{datasets}
