% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\docType{data}
\name{mortality}
\alias{mortality}
\title{Sex-specific instaneous mortality of American shad}
\format{
A data frame with 6 observations of 3 variables:
\describe{

\code{region}{ Life-history region}

\code{sex}{ Fish sex (gender)}

\code{M}{ Instantaneous natural mortality}
}
}
\source{
Atlantic States Marine Fisheries Commission
}
\usage{
mortality
}
\description{
A dataset containing sex-specific instantaneous
mortality estimates for each life-history region 
(NI = Northern iteroparous, SI = Southern iteroparous, 
SP = Semelparous)
}
\keyword{datasets}
