% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_rh.R
\docType{data}
\name{mortality_rh}
\alias{mortality_rh}
\title{Sex-specific instantaneous natural mortality of river herring}
\format{
A data frame with 300 observations of 5 variables:
\describe{

\code{region}{ Genetic reporting group region}

\code{sex}{ Sex of fish}

\code{age}{ Age in years}

\code{M}{ Instantaneous natural mortality}

\code{species}{ Species}
}
}
\source{
Atlantic States Marine Fisheries Commission
}
\usage{
mortality_rh
}
\description{
A dataset containing sex- and age-specific instantaneous
natural mortality estimates for each species and genetic reporting region
}
\keyword{datasets}
