% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olney_mcbride.R
\docType{data}
\name{olney_mcbride}
\alias{olney_mcbride}
\title{Weight-batch fecundity relationships for American Shad}
\format{A data frame with 3 observations of 4 variables:
\describe{

\code{system}{ Study system}

\code{region}{ Life-history region: NI = Northern iteroparous,
SI = Southern iteroparous, SP = Semelparous}

\code{alpha}{ Intercept}

\code{beta}{ Slope}
}}
\source{
Olney and McBride (2003)
}
\usage{
olney_mcbride
}
\description{
A dataset containing weight-batch fecundity relationships
for St. Johns, York, and Connecticut Rivers from Olney and McBride (2003).
Regressions fit log10 batch fecundity as a function of
log10 body weight (g).
}
\keyword{datasets}
