% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrunc_norm.R
\name{rtrunc_norm}
\alias{rtrunc_norm}
\title{Simulate random samples from truncated normal distribution}
\usage{
rtrunc_norm(n, a = -Inf, b = Inf, mean, sd)
}
\arguments{
\item{n}{Number of random samples}

\item{a}{Lower bound(s)}

\item{b}{Upper bound(s)}

\item{mean}{Mean(s)}

\item{sd}{Standard deviation(s)}
}
\description{
Simulate random draws from a truncated normal distribution based
using inverse cumulative distribution function for normal distribution.
}
\details{
This approach is an approximation based on the normal distribution
for case application to tested uses in the `anadrofish` package. It should
not be considered a generalized solution. This function is used to replace
the \code{rtruncnorm} function from the \code{truncnorm} package to eliminate
need for compilation in the \code{anadrofish} package and eliminate
dependency on \code{truncnorm} (which requires compilation).
}
\examples{
rtrunc_norm(n = 1, a = 0, b = Inf, mean = 6.1, sd = 0.5)

}
\references{
Answers by Glen_b and JohannesNE to this post on Cross Validated:
https://stats.stackexchange.com/questions/56747/simulate-constrained-normal-on-lower-or-upper-bound-in-r
were helpful.
}
