% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anadrofish.R
\docType{package}
\name{shadia}
\alias{shadia}
\alias{shadia-package}
\title{anadrofish: Anadromous Fish Population Responses to Habitat Changes}
\description{
The anadrofish package provides tools for running
coastwide or river-specific population models for migratory
fish lacking fine-scale information about intracatchment
homing.
}
\section{Functions called directly}{

 \describe{
   \code{\link{beverton_holt}} \cr
   \code{\link{lower95}} \cr
   \code{\link{make_fec}} \cr
   \code{\link{make_pop}} \cr
   \code{\link{make_spawners}} \cr
   \code{\link{upper95}} \cr
 }
}

\section{Data}{

  \describe{
    \code{\link{habitat}} \cr
  }
}

