% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop_shiny.R
\name{sim_pop_shiny}
\alias{sim_pop_shiny}
\title{Launch Anadromous Fish Population Simulator}
\usage{
sim_pop_shiny()
}
\value{
A Shiny application
}
\description{
Interactive interface for running anadrofish population simulations
}
