% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper95.R
\name{upper95}
\alias{upper95}
\title{Calculate upper 95 \% confidence limit}
\usage{
upper95(x)
}
\arguments{
\item{x}{A numeric vector or matrix.}
}
\value{
Numeric vector of length 1 containing 95% UCL for \code{x}.
}
\description{
Function with a short name to calculate 97.5th percentile
of a distribution. Calls \code{stats::quantile}, 
but is shorter and is easier to use in calls to \code{apply} family
of functions.
}
\examples{
# Example usage
\dontrun{
  
 x <- rnorm(1e4, 0, 1)
 
 upper95(x)
  
}
}
