% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbgf_ale.R
\docType{data}
\name{vbgf_ale}
\alias{vbgf_ale}
\title{Sex-specific von Bertalanffy growth parameters for alewife}
\format{
A data frame with 1,116,000 observations of 5 variables:
\describe{

\code{Region}{ Genetic reporting group}

\code{Sex}{ Sex of fish}

\code{Linf}{ Asymptotic length of fish}

\code{K}{ Brody growth coefficient}

\code{t0}{ Size at age zero}

}
}
\source{
ASMFC (2024)
}
\usage{
vbgf_ale
}
\description{
A dataset containing sex-specific estimates of von Bertalanffy
growth parameters for alewife from the 2024 ASMFC stock assessment.
}
\keyword{datasets}
