% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbgf_NI.R
\docType{data}
\name{vbgf_NI}
\alias{vbgf_NI}
\title{Sex-aggregate von Bertalanffy growth parameters for 
Northern Iteroparous American shad}
\format{A data frame with 1500 observations of 3 variables:
\describe{

\code{Linf}{ Asymptotic length of fish}

\code{K}{ Brody growth coefficient}

\code{t0}{ Size at age zero}
}}
\source{
Gilligan, E. K. 2019. Understanding influences of climate change
on American Shad recovery, fisheries management, and dam impacts. Master's
Thesis, SUNY Oneonta, Oneonta, NY, USA.
}
\usage{
vbgf_NI
}
\description{
A dataset containing sex-aggregate estimates
of von Bertalanffy growth parameters for Northern 
Iteroparous life-history region. Data frame contains posterior
draws from coast-wide growth models (Gilligan 2019).
}
\keyword{datasets}
