% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbgf_SP.R
\docType{data}
\name{vbgf_SP}
\alias{vbgf_SP}
\title{Sex-aggregate von Bertalanffy growth parameters for
Semelparous American shad}
\format{
A data frame with 1500 observations of 3 variables:
\describe{
    \item{\code{Linf}}{Asymptotic length of fish}
    \item{\code{K}}{Brody growth coefficient}
    \item{\code{t0}}{Size at age zero}
}
}
\source{
Gilligan, E. K. 2019. Understanding influences of climate change
on American Shad recovery, fisheries management, and dam impacts. Master's
Thesis, SUNY Oneonta, Oneonta, NY, USA.
}
\usage{
vbgf_SP
}
\description{
A dataset containing sex-aggregate estimates
of von Bertalanffy growth parameters for Semelparous
life-history region. Data frame contains posterior
draws from coast-wide growth models (Gilligan 2019).
}
\keyword{datasets}
