% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_unequal_vectors.R
\name{add_unequal_vectors}
\alias{add_unequal_vectors}
\title{Add vectors of unequal length}
\usage{
add_unequal_vectors(x, y)
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}
}
\value{
A vector of length one containing pre-spawn survival.
}
\description{
Function used to add aligned vectors of unequal length. Used
to combine male and female cohorts when max_ages are not equal.
}
\references{
https://stackoverflow.com/questions/2307443/how-to-add-two-vectors-without-repeating-in-r
}
