% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_habitat_template.R
\name{custom_habitat_template}
\alias{custom_habitat_template}
\title{Make custom habitat for an existing river or a template for a new 
river of interest}
\usage{
custom_habitat_template(
  species = c("AMS", "ALE", "BBH"),
  built_in = TRUE,
  river
)
}
\arguments{
\item{species}{Species for which population dynamics will be simulated.
Choices include American shad (\code{"AMS"}), alewife (\code{"ALE"}), and
blueback herring (\code{"BBH"}).}

\item{built_in}{A logical indicating whether the custom habitat template is a
a subset of built-in habitat datasets for the \code{species} indicated.}

\item{river}{Character string specifying river name or NULL for custom river.
If making custom habitat from an existing river in \code{\link{habitat}} (American
shad), \code{\link{habitat_ale}} (Alewife), or \code{\link{habitat_bbh}} (Blueback herring), 
then river must be included in rivers from \code{\link{get_rivers}} for the target 
species. Otherwise, an arbitrary character string identifying river name 
is acceptable.}
}
\description{
Function used to make habitat for rivers listed
in \code{\link{get_rivers}} from the built-in dataset(s)
}
