% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_output.R
\name{fill_output}
\alias{fill_output}
\title{Fill output vectors}
\usage{
fill_output(.sim_pop, sex_specific = FALSE)
}
\arguments{
\item{.sim_pop}{A hidden environment in the calling frame 
of \code{\link{sim_pop}}. Arbitrarily any named object with
names matching those used in the function.}

\item{sex_specific}{Logical inherited from \code{\link{sim_pop}}
indicating whether to use sex-specific output.}
}
\value{
list of output vectors
}
\description{
Internal function to fill output vectors for default
arguments in \code{sim_pop} and the output (\code{spawners}).

For internal use in \code{\link{sim_pop}}. Not intended 
to be called directly.
}
\keyword{Internal}
