% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_output.R
\name{fill_output}
\alias{fill_output}
\title{Fill output vectors}
\usage{
fill_output()
}
\value{
list of output vectors
}
\description{
Internal function to fill output vectors for default
arguments in \code{sim_pop} and the output (\code{spawners}).

Not intended to be called directly.
}
