% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_govt.R
\name{get_govt}
\alias{get_govt}
\title{Get governmental unit for specified river by species}
\usage{
get_govt(river, species = c("ALE", "AMS", "BBH"), custom_habitat = NULL)
}
\arguments{
\item{river}{Character string specifying river name}

\item{species}{Character string specifying species
Choices include American shad (\code{"AMS"}), alewife (\code{"ALE"}), and
blueback herring (\code{"BBH"}).}

\item{custom_habitat}{A dataframe containing columns corresponding to the
those in the output from custom_habitat_template(). NEED TO ADD LINK.}
}
\value{
A character string indicating governmental district corresponding
to river
}
\description{
Function used to get governmental unit for rivers listed
in output of \code{\link{get_rivers}} from the built-in 
habitat data sets
}
