% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region.R
\name{get_region}
\alias{get_region}
\title{Get region for specified river by species}
\usage{
get_region(river, species = c("ALE", "AMS", "BBH"))
}
\arguments{
\item{river}{Character string specifying river name}

\item{species}{Character string specifying species for simulation.
Choices include American shad (\code{"AMS"}), alewife (\code{"ALE"}), and
blueback herring (\code{"BBH"}).}
}
\value{
a data.frame with 4 variables containing dam name,
latitude and longitude, and dam order in the watershed
}
\description{
Function used to get region for rivers listed
in output of \code{\link{get_rivers}} from the built-in 
habitat data sets
}
