% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rivers.R
\name{get_rivers}
\alias{get_rivers}
\title{Rivers included in package}
\usage{
get_rivers()
}
\value{
A character vector of rivers that are included in both 
\link{inventory} and \link{habitat}.
}
\description{
Get rivers included in package. A function with no arguments.
}
\details{
The primary uses of this function are (1) so the user can see 
rivers that can be used for simulation, and (2) so the output can be passed 
to a random sampler for use in stochastic simulation via parallel processing.
}
