% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\docType{data}
\name{habitat}
\alias{habitat}
\title{American shad habitat}
\format{
A data frame with 868 observations of 22 variables:
\describe{
    \item{\code{TERMCODE}}{Code used to identify regional, state, or river units in which features are located, assists with subselecting features for analysis.}
    \item{\code{system}}{Name or name(s) of rivers included in catchment.}
    \item{\code{region}}{Life-history region within which river is located (NI = northern iteroparous, SI = southern iteroparous, SP = semelparous).}
    \item{\code{huc_code}}{n-digit huc code for the huc identifier used in \code{TERMCODE}.}
    \item{\code{huc_level}}{Watershed level used to determine \code{huc_code}.}
    \item{\code{count}}{Number of habitat segments (features) included within \code{system}.}
    \item{\code{total}}{Total historical habitat available in \code{system} (sq km).}
    \item{\code{UNIQUE_ID}}{Unique value given to each dam and outlet feature.}
    \item{\code{type}}{Type of feature, either \code{dam} or \code{outlet}.}
    \item{\code{catchmentID}}{Unique value of the coastal outlet in which the dam is located upstream.}
    \item{\code{habitat_sqkm}}{Total habitat upstream, given in square kilometers of river area.}
    \item{\code{habitatSegment_sqkm}}{Amount of habitat in the segment upstream of the feature, between the feature and the next group of adjacent upstream features.}
    \item{\code{functional_habitatSegment_sqkm}}{The value of \code{habitatSegment_sqkm} multiplied by the compounded passage fraction from outlet up to the feature.}
    \item{\code{PassageToHabitat}}{Passage fraction at feature, compounded over this and all downstream features. Default dam passage is zero, default outlet passage is one. These values can be modified at start of model run.}
    \item{\code{terminal_name_huc10}}{Name of HUC10 watershed of feature's outlet.}
    \item{\code{terminal_name_huc8}}{Name of HUC8 watershed of feature's outlet.}
    \item{\code{terminal_name_huc6}}{Name of HUC6 watershed of feature's outlet.}
    \item{\code{terminal_name_huc4}}{Name of HUC4 watershed of feature's outlet.}
    \item{\code{dam_name}}{Name of dam feature, "n/a" if \code{outlet}.}
    \item{\code{latitude}}{Latitudinal coordinate of feature centroid.}
    \item{\code{longitude}}{Longitudinal coordinate of feature centroid.}
    \item{\code{dam_order}}{Order in which dam is located in river, starting from coastal outlet. Outlets have code = 0.}
}
}
\source{
Samuel G. Roy: \url{https://github.com/samGroy/AmShadProject}
}
\usage{
habitat
}
\description{
A dataset containing the surface area of American
shad habitat segments by feature (dam or outlet) in
Atlantic coast drainages of North America.
}
\references{
Zydlewski, J., D. S. Stich, S. Roy, M. Bailey, T. Sheehan, and
K. Sprankle. 2021. What Have We Lost? Modeling Dam Impacts on American Shad
Populations Through Their Native Range. Frontiers in Marine Science 8.
DOI: https://doi.org/10.3389/fmars.2021.734213.
}
\keyword{datasets}
