% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\docType{data}
\name{habitat}
\alias{habitat}
\title{American shad habitat}
\format{A data frame with 4216 observations of 16 variables:
\describe{
\code{UNIQUE_ID}{ unique value given to each dam and outlet feature.}

\code{type}{ type of feature, either\code{dam} or \code{outlet}.}

\code{catchmentID}{ unique value of the coastal outlet in which the dam is located upstream.}

\code{habitat_sqkm}{ total habitat upstream, given in square kilometers of river area.}

\code{habitatSegment_sqkm}{ amount of habitat in the segment upstream of the feature, between the feature and the next group of adjacent upstream features.}

\code{functional_habitatSegment_sqkm}{ the value of \code{habitatSegment_sqkm} multiplied by the compounded passage fraction from outlet up to the feature.}

\code{PassageToHabitat}{ passage fraction at feature, compounded over this and all downstream features. Default dam passage is zero, default outlet passage is one. These values can be modified at start of model run.}

\code{TERMCODE}{ code used to identify regional, state, or river units in which features are located, assists with subselecting features for analysis.}

\code{terminal_name_huc10}{ name of HUC10 watershed of feature's outlet.}

\code{terminal_name_huc8}{ name of HUC8 watershed of feature's outlet.}

\code{terminal_name_huc6}{ name of HUC6 watershed of feature's outlet.}

\code{terminal_name_huc4}{ name of HUC4 watershed of feature's outlet.}

\code{dam_name}{ name of dam feature, "n/a" if \code{outlet}.}

\code{latitude/longitude}{ coordinates of feature.}

\code{dam_order}{ order in which dam is located in river, starting from coastal outlet. Outlets have code = 0.}
}}
\source{
Samuel G. Roy: \url{https://github.com/samGroy/AmShadProject}
}
\usage{
habitat
}
\description{
A dataset containing the surface area of American 
shad habitat segments by feature (dam or outlet) in
Atlantic coast drainages.
}
\keyword{datasets}
