% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_ale.R
\docType{data}
\name{habitat_ale}
\alias{habitat_ale}
\title{Alewife habitat}
\format{
A data frame with 246,259 observations of 26 variables:
\describe{

\code{Names}{ NEED TO ADD DESCRIPTIONS}
}
}
\source{
Shawn Snyder
}
\usage{
habitat_ale
}
\description{
A dataset containing the surface area of Alewife 
habitat segments by feature (dam or outlet) in
Atlantic coast drainages.
}
\examples{
print(habitat_ale)



}
\keyword{datasets}
