% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_bbh.R
\docType{data}
\name{habitat_bbh}
\alias{habitat_bbh}
\title{Blueback herring habitat}
\format{
A data frame with 31859 observations of 24 variables:
\describe{

\code{Names}{ NEED TO ADD DESCRIPTIONS}

}
}
\source{
Shawn Snyder
}
\usage{
habitat_bbh
}
\description{
A dataset containing the surface area of blueback herring 
habitat segments by feature (dam or outlet) in
Atlantic coast drainages.
}
\keyword{datasets}
