% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_downstream.R
\name{make_downstream}
\alias{make_downstream}
\title{Calculate downstream survival given dam passage scenario.}
\usage{
make_downstream(habitat_data, river, downstream, upstream)
}
\arguments{
\item{habitat_data}{An internal data.frame in \code{\link{sim_pop}} 
containing habitat estimates (km2) for the scenario and region selected.}

\item{river}{Character string specifying river name.}

\item{downstream}{Numeric indicating proportional downstream survival 
through a single dam.}

\item{upstream}{Numeric indicating proportional upstream passage 
through a single dam.}
}
\value{
Numeric of length 1 representing catchment-scale 
downstream migration mortality for juvenile or adult fish.
}
\description{
Function used to create population-level survival during
out-migration through dams if downstream passage < 1.
}
\details{
This function assigns cumulative downstream passage values 
to all features in \code{habitat_data} corresponding to \code{river}.
It then calculates the proportion of habitat in each 
habitat segment, and weights downstream mortality at the catchment-scale
by proportion of habitat. This implicitly assumes that fish are distributed 
throughout the river during spawning in proportion to available
habitat.
}
