% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eggs.R
\name{make_eggs}
\alias{make_eggs}
\title{Simulate number of eggs per fish}
\usage{
make_eggs(region, max_age)
}
\arguments{
\item{region}{The life-history region for the population.}

\item{max_age}{The maximum age of fish in the population(s). A numeric vector of length 1.}
}
\value{
A vector containing age-specific potential annual 
fecundity with \code{length = max_age}.
}
\description{
Function used to simulate number of eggs 
produced per female (potential annual fecundity, PAF).
}
\references{
Lehman, B. A. 1953. Fecundity of Hudson River shad. 
US Fish and Wildlife Service, Research Reports 33:1–8.
}
