% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eggs.R
\name{make_eggs}
\alias{make_eggs}
\title{Simulate number of eggs per fish}
\usage{
make_eggs(region, max_age, egg_est)
}
\arguments{
\item{region}{The life-history region for the population.}

\item{max_age}{The maximum age of fish in the population(s).
A numeric vector of length 1.}

\item{egg_est}{The method used to estimate number of eggs
per female in each age class. The default \code{'Olney_McBride'}
uses predictive equations from Olney and McBride (2003) to estimate
batch fecundity from weight-fecundity relationships
using coefficients in \code{Olney} for the life-history region
corresponding to the specified river. Number of batches for this
approach is currently drawn from a random normal distribution
with a mean of 6.7 and a standard deviation of 2.1 (McBride et al. 2016).

\code{Lehman} uses an alternative equation from Lehman (1953)
to predict mean fecundity from length.}
}
\value{
A vector containing age-specific potential annual 
fecundity with \code{length = max_age}.
}
\description{
Function used to simulate number of eggs 
produced per female (potential annual fecundity, PAF).
}
\references{
Lehman, B. A. 1953. Fecundity of Hudson River shad. 
US Fish and Wildlife Service, Research Reports 33:1–8.
}
