% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eggs.R
\name{make_eggs}
\alias{make_eggs}
\title{Simulate number of eggs per fish}
\usage{
make_eggs(river)
}
\arguments{
\item{river}{The river for which eggs will be simulated.}
}
\value{
A vector containing age-specific potential annual 
fecundity with \code{length = max_age}.
}
\description{
Function used to simulate number of eggs 
produced per female (potential annual fecundity, PAF),
for rivers included in \code{\link{get_rivers}}.
}
\section{Details}{

The default uses predictive equations from Olney and McBride (2003) 
to simulate batch fecundity from weight-fecundity relationships
using coefficients in \code{\link{olney_mcbride}} for the life-history region
corresponding to the specified river. Number of batches for this
approach is currently drawn from a random normal distribution
with a mean of 6.7 and a standard deviation of 2.1 (McBride et al. 2016).
}

\examples{
make_eggs(river="Susquehanna")

}
\references{
Olney, J. E. and R. S. McBride. 2003. Intraspecific 
variation in batch fecundity of American shad (Alosa sapidissima): 
revisiting the paradigm of reciprocal latitudinal trends 
in reproductive traits. American Fisheries Society 
Symposium 35:185-192.

McBride, R. S., R. Ferreri, E. K. Towle, J. M. Boucher, and 
G. Basilone, G. 2016. Yolked oocyte dynamics support agreement 
between determinate- and indeterminate-method estimates of annual 
fecundity for a northeastern United States population of 
American shad. PLoS ONE 11(10): 10.1371/journal.pone.0164203
}
