% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eggs_rh.R
\name{make_eggs_rh}
\alias{make_eggs_rh}
\title{Simulate number of eggs per fish for river herring}
\usage{
make_eggs_rh(river, species = c("ALE", "BBH"), custom_habitat = NULL)
}
\arguments{
\item{river}{The river for which eggs will be simulated.}

\item{species}{Species for which population dynamics will be simulated.
Choices include alewife (\code{"ALE"}), and blueback herring (\code{"BBH"}).}

\item{custom_habitat}{A dataframe containing columns corresponding to the
those in the output from \code{\link{custom_habitat_template}}. The default,
\code{NULL} uses the default habitat data set for a given combination of
\code{species} and \code{river}.}
}
\value{
A vector containing age-specific potential annual 
fecundity with \code{length = max_age}.
}
\description{
Function used to simulate number of eggs 
produced per female (total fecundity) for river herring in
rivers included in \code{\link{get_rivers}}.
}
\section{Details}{
 
The default methods of alewife and blueback herring uses a stochastic,
MCMC sampling approach to draw correlated sets of von Bertalanffy growth
parameters from built-in \code{vbgf_...} datasets, 
length-weight regression parameters in \code{link{lw_pars_rh}}, and
length-fecundity relationships from Sullivan et al. (2019) to simulate the
number of eggs per female within genetic reporting groups (\code{region}).
}

\examples{
make_eggs_rh(river = "Upper Susquehanna", species = "BBH")

}
\references{
Atlantic States Marine Fisheries Commission. 2024. River herring
benchmark stock assessment and peer-review report. ASMFC, Arlington, VA. 
URL: https://asmfc.org/uploads/file/66f59e40RiverHerringAssessment_PeerReviewReport_2024.pdf

Sullivan, K.M, M.M. Bailey, and D.L. Berlinksky. 2019. Digital Image Analysis
as a Technique for Alewife Fecundity Estimation in a New Hampshire River. 
North American Journal of Fisheries Management 39:353-361.
}
