% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_habitat.R
\name{make_habitat}
\alias{make_habitat}
\title{Make cumulative habitat for river}
\usage{
make_habitat(river, upstream, historical = FALSE)
}
\arguments{
\item{river}{Character string specifying river name}

\item{upstream}{Proportional upstream passage through dams.}

\item{habitat_data}{A built-in data.frame containing habitat 
estimates (km2) for the scenario and region selected.}
}
\description{
Function used to make habitat for rivers listed
in \code{\link{get_rivers}} from the built-in 
dataset(s)
}
