% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_habitat.R
\name{make_habitat}
\alias{make_habitat}
\title{Make cumulative habitat for river}
\usage{
make_habitat(river, upstream, historical)
}
\arguments{
\item{river}{Character string specifying river name}

\item{upstream}{Proportional upstream passage through dams.}

\item{historical}{TESTING PARAM ONLY. Logical indicating whether this
is an historical analysis. Default is FALSE. If TRUE, then use test 
habitat data for Merrimack, Presumpscot, or Salmon Falls rivers. Not 
implemented for any other systems.}
}
\description{
Function used to make habitat for rivers listed
in \code{\link{get_rivers}} from the built-in 
dataset(s)
}
