% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_habitat.R
\name{make_habitat}
\alias{make_habitat}
\title{Make cumulative habitat for river}
\usage{
make_habitat(
  river,
  species = c("AMS", "ALE", "BBH"),
  upstream,
  custom_habitat = NULL
)
}
\arguments{
\item{river}{Character string specifying river name}

\item{species}{Species for which population dynamics will be simulated.
Choices include American shad (\code{"AMS"}), alewife (\code{"ALE"}), and
blueback herring (\code{"BBH"}).}

\item{upstream}{Proportional upstream passage through dams.}

\item{custom_habitat}{A dataframe containing columns corresponding to the
those in the output from custom_habitat_template(). NEED TO ADD LINK.}
}
\description{
Function used to make habitat for rivers listed
in \code{\link{get_rivers}} from the built-in dataset(s)
}
