% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_habitat.R
\name{make_habitat}
\alias{make_habitat}
\title{Project population}
\usage{
make_habitat(habitat_data, river, type)
}
\arguments{
\item{habitat_data}{A built-in \code{data.frame} containing habitat 
estimates (km2) for the scenario and region selected.}

\item{river}{Character string specifying river name}

\item{type}{Character string indicating the type of habitat 
to be considered. Possible values include \code{functional}, 
or \code{segment}. For \code{PassageToHabitat < 1}, 
\code{type = 'functional'} adjusts \code{habitatSegment_sqkm} 
proportionally and uses the variable \code{functional_habitatSegment_sqkm} 
for calculations. See \code{?habitat} for explanation of built-in 
data sets.}
}
\description{
Function used to make habitat from the built-in 
dataset(s)
}
