% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_habitat.R
\name{make_habitat}
\alias{make_habitat}
\title{Project population}
\usage{
make_habitat(habitat_data, river, upstream)
}
\arguments{
\item{habitat_data}{A built-in data.frame containing habitat 
estimates (km2) for the scenario and region selected.}

\item{river}{Character string specifying river name}

\item{upstream}{Proportional upstream passage through dams.}
}
\description{
Function used to make habitat for rivers listed
in \code{\link{get_rivers}} from the built-in 
dataset(s)
}
