% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lat.R
\name{make_lat}
\alias{make_lat}
\title{Retrieve latitude for American shad rivers.}
\usage{
make_lat(river, species = c("AMS", "ALE", "BBH"))
}
\arguments{
\item{river}{Character string specifying river name. See 
\code{\link{get_rivers}}.}
}
\description{
Function used to extract river-specific latitude
from built-in data sets for American shad. Not implemented for river herring.
}
\examples{
make_lat(river = 'Susquehanna')

}
