% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lat.R
\name{make_lat}
\alias{make_lat}
\title{Retrieve latitude for shad rivers.}
\usage{
make_lat(river)
}
\arguments{
\item{river}{Character string specifying river name. See 
\code{\link{get_rivers}}.}
}
\description{
Function used to extract river-specific latitude
from built-in data sets.
}
\examples{
make_lat(river = 'Susquehanna')

}
