% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_maxage.R
\name{make_maxage}
\alias{make_maxage}
\title{Make maximum age for population using built-in data sets.}
\usage{
make_maxage(river, sex = c("female", "male"))
}
\arguments{
\item{river}{River for which maximum age is needed.}

\item{sex}{Sex of fish. If not specified, then mean of
male and female maximum ages is returned.}
}
\value{
A numeric vector of \code{length = 1} containing
maximum age of fish in population.
}
\description{
The purpose of this function is to query the maximum age 
in the selected river from built-in data set containing 
region-specific \code{\link{max_ages}} and.
}
\examples{
make_maxage(river = "Susquehanna")

}
\references{
Atlantic States Marine Fisheries Commission
}
