% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_maxage.R
\name{make_maxage}
\alias{make_maxage}
\title{Make maximum age for population using built-in data sets.}
\source{
Atlantic States Marine Fisheries Commission
}
\usage{
make_maxage(
  river,
  sex = c("female", "male"),
  species = c("AMS", "ALE", "BBH"),
  custom_habitat = NULL
)
}
\arguments{
\item{river}{River for which maximum age is needed.}

\item{sex}{Sex of fish. If not specified, then mean of
male and female maximum ages is returned for American shad,
or pooled sex data used for river herring.}

\item{species}{Species for which population dynamics will be simulated.
Choices include American shad (\code{"AMS"}), alewife (\code{"ALE"}), and
blueback herring (\code{"BBH"}).}

\item{custom_habitat}{A dataframe containing columns corresponding to the
those in the output from \code{\link{custom_habitat_template}}. The default,
\code{NULL} uses the default habitat data set for a given combination of
\code{species} and \code{river}.}
}
\value{
A numeric vector of \code{length = 1} containing
maximum age of fish in population.
}
\description{
The purpose of this function is to query the maximum age 
in the selected river from built-in data set containing 
region-specific \code{\link{max_ages}} and.
}
\examples{
make_maxage(river = "Susquehanna", species = "AMS")

}
\references{
Atlantic States Marine Fisheries Commission (ASMFC). 2020. 
American shad benchmark stock assessment and peer-review report. ASMFC, 
Arlington, VA.

Atlantic States Marine Fisheries Commission. 2024. River herring
benchmark stock assessment and peer-review report. ASMFC, Arlington, VA. 
URL: https://asmfc.org/uploads/file/66f59e40RiverHerringAssessment_PeerReviewReport_2024.pdf
}
