% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mortality.R
\name{make_mortality}
\alias{make_mortality}
\title{Make instantaneous natural mortality.}
\usage{
make_mortality(
  river,
  sex = c(NULL, "female", "male"),
  species = c("AMS", "ALE", "BBH"),
  custom_habitat = NULL
)
}
\arguments{
\item{river}{River for which maximum age is needed.}

\item{sex}{Fish sex. If \code{NULL} (default) or \code{"Pooled"}, then mean
of male and female estimates is used.}

\item{species}{Species for which population dynamics will be simulated.
Choices include American shad (\code{"AMS"}), alewife (\code{"ALE"}), and
blueback herring (\code{"BBH"}).}

\item{custom_habitat}{A dataframe containing columns corresponding to the
those in the output from custom_habitat_template(). NEED TO ADD LINK.}
}
\value{
A numeric vector of \code{length = 1} containing
natural instantaneous mortality of fish in population.
}
\description{
The purpose of this function is to query the 
instantaneous natural mortality (regional estimates) 
for the selected river from built-in data set containing 
region-specific \code{\link{mortality}} and.
}
\examples{
make_mortality(river = "Penobscot", species = "BBH")

}
\references{
Atlantic States Marine Fisheries Commission
}
