% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mortality.R
\name{make_mortality}
\alias{make_mortality}
\title{Make instantaneous natural mortality.}
\usage{
make_mortality(river, sex = c(NULL, "female", "male"))
}
\arguments{
\item{river}{River for which maximum age is needed.}

\item{sex}{Fish sex. If \code{NULL} (default), then mean
of male and female estimates is used.}
}
\value{
A numeric vector of \code{length = 1} containing
natural instantaneous mortality of fish in population.
}
\description{
The purpose of this function is to query the 
instantaneous natural mortality (regional estimates) 
for the selected river from built-in data set containing 
region-specific \code{\link{mortality}} and.
}
\examples{
make_mortality(river = "Susquehanna")

}
\references{
Atlantic States Marine Fisheries Commission
}
