% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mortality.R
\name{make_mortality}
\alias{make_mortality}
\title{Make instantaneous natural mortality.}
\source{
Atlantic States Marine Fisheries Commission
}
\usage{
make_mortality(
  river,
  sex = c(NULL, "female", "male"),
  species = c("AMS", "ALE", "BBH", "EEL"),
  max_age,
  custom_habitat = NULL
)
}
\arguments{
\item{river}{River for which maximum age is needed.}

\item{sex}{Fish sex. If \code{NULL} (default) or \code{"Pooled"}, then mean
of male and female estimates is used.}

\item{species}{Species for which population dynamics will be simulated.
Choices include American shad (\code{"AMS"}), alewife (\code{"ALE"}),
blueback herring (\code{"BBH"}), and American eel (\code{EEL}).}

\item{custom_habitat}{A dataframe containing columns corresponding to the
those in the output from \code{\link{custom_habitat_template}}. The default,
\code{NULL} uses the default habitat data set for a given combination of
\code{species} and \code{river}.}
}
\value{
A numeric vector of \code{length = 1} containing
natural instantaneous mortality of fish in population.
}
\description{
The purpose of this function is to query the
instantaneous natural mortality (regional estimates)
for the selected river from built-in data set containing
region-specific \code{\link{mortality}} for American shad or
\code{\link{mortality_rh}} for river herring.
}
\examples{
make_mortality(river = "Penobscot", species = "BBH")

}
\references{
Atlantic States Marine Fisheries Commission (ASMFC). 2020.
American shad benchmark stock assessment and peer-review report. ASMFC,
Arlington, VA.

Atlantic States Marine Fisheries Commission. 2024. River herring
benchmark stock assessment and peer-review report. ASMFC, Arlington, VA.
URL: https://asmfc.org/uploads/file/66f59e40RiverHerringAssessment_PeerReviewReport_2024.pdf

Atlantic States Marine Fisheries Commission. 2012. American Eel benchmark
stock assessment. ASMFC, Arlington, VA.
URL: https://asmfc.org/wp-content/uploads/2024/11/americanEelBenchmarkStockAssessmentReport_May2012.pdf
}
