% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_output.R
\name{make_output}
\alias{make_output}
\title{Make output vectors}
\usage{
make_output(nyears, sex_specific = FALSE)
}
\arguments{
\item{nyears}{Number of years for simulation output.}

\item{sex_specific}{Logical inherited from \code{\link{sim_pop}}
indicating whether to use sex-specific output.}
}
\description{
Internal function to create output
vectors in \code{\link{sim_pop}}. Not intended to be called directly.
}
\keyword{Internal}
