% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_output.R
\name{make_output}
\alias{make_output}
\title{Make output vectors}
\usage{
make_output(nyears)
}
\arguments{
\item{nyears}{Number of years for simulation output.

Not intended to be called directly.}
}
\description{
Internal function to create output 
vectors in \code{\link{sim_pop}}.
}
