% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_output_row.R
\name{make_output_row}
\alias{make_output_row}
\title{Make one row of simulation output}
\usage{
make_output_row(.sim_pop, sex_specific = FALSE)
}
\arguments{
\item{.sim_pop}{The simulation environment.}

\item{sex_specific}{Logical indicating whether to use sex-specific output.}
}
\value{
A named list representing one row of output.
}
\description{
Internal function used to capture all output fields from
\code{.sim_pop} for a single year. Called inside the simulation loop
in \code{\link{sim_pop}}.
}
\keyword{Internal}
