% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pop.R
\name{make_pop}
\alias{make_pop}
\title{Make population}
\usage{
make_pop(max_age, nM, fM, n_init)
}
\arguments{
\item{max_age}{The maximum age of fish in the population(s). A numeric vector of length 1.}

\item{nM}{Instantaneous natural mortality rate. 
A numeric vector of length 1.}

\item{fM}{Instantaneous fishing mortality rate. 
A numeric vector of length 1.}

\item{n_init}{Initial population abundance (includes all age classes).}
}
\value{
A vector containing age-specific abundances 
with \code{length = max_age}.
}
\description{
Function used to seed initial age-structured
population for simulation based on life-history
characteristics. Uses an initial population seed, mortality
estimates, and maximum age to create a starting population.
}
\examples{
# Example usage
\dontrun{
  
# Simulate a fish population at stable age distribution
  pop = make_pop(max_age, nM, fM, n_init)
  
}
}
