% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pop.R
\name{make_pop}
\alias{make_pop}
\title{make_pop}
\usage{
make_pop(max_age = 11, nM = 0.38, fM = 0, n_init = 1e+05)
}
\arguments{
\item{max_age}{The maximum age of fish in the population(s). A numeric vector of length 1.}

\item{nM}{Instantaneous natural mortality rate. A numeric vector of length 1.}

\item{fM}{Instantaneous fishing mortality rate. A numeric vector of length 1.}

\item{n_init}{Initial population abundance (includes all age classes)}
}
\value{
A vector containing age-specific abundances. A numeric vector with length = max_age.
}
\description{
Function used to seed initial
populations for simulation based on life-history
characteristics and stable age distributions
}
\references{
None yet.
}
