% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pop.R
\name{make_pop}
\alias{make_pop}
\title{Make population}
\usage{
make_pop(species, max_age, nM, fM, n_init)
}
\arguments{
\item{species}{Species for which population dynamics will be simulated.
Choices include American shad (\code{"AMS"}), alewife (\code{"ALE"}), 
blueback herring (\code{"BBH"}), and American eel (\code{"EEL"}).}

\item{max_age}{The maximum age of fish in the population(s).
A numeric vector of length 1.}

\item{nM}{Instantaneous natural mortality rate.
A numeric vector of length for (\code{"AMS"} or a vector of
length \code{max_age} for \code{"ALE"}, \code{"BBH"}, and (\code{"EEL"}).}

\item{fM}{Instantaneous fishing mortality rate.
A numeric vector of length 1.}

\item{n_init}{Initial population abundance (includes all age classes).}
}
\value{
A vector containing age-specific abundances
with \code{length = max_age}.
}
\description{
Function used to seed initial age-structured
population for simulation based on life-history
characteristics. Uses an initial population seed, mortality
estimates, and maximum age to create a starting population.
}
\examples{
# Example usage
\dontrun{
  
# Simulate a fish population at stable age distribution
pop <- make_pop(max_age, nM, fM, n_init)
  
}
}
