% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_postspawn.R
\name{make_postspawn}
\alias{make_postspawn}
\title{Post-spawning survival}
\usage{
make_postspawn(iteroparity, nM)
}
\arguments{
\item{iteroparity}{Rate of iteroparity.}

\item{nM}{Instantaneous annual mortality.}
}
\description{
Function used to estimate post-spawn survival 
from proportion of repeat spawners by latitude (Leggett and
Cascardden 1978, Bailey and Zydlewski 2013) and natural mortality
by life-history region.
}
\references{
Bailey, M.M., and J. D. Zydlewski. 2013. To stock or not
to stock? Assessing the restoration potential of a remnant 
American shad spawning run with hatchery supplementation. North 
American Journal of Fisheries Management 33:459–467.

Leggett, W., and J. E. Cascardden. 1978. Latitudinal Variation in 
Reproductive Characteristics of American Shad (Alosa sapidissima): 
Evidence for Population Specific Life History Strategies in Fish.
Journal of the Fisheries Research Board of Canada 35:1469-1478.
}
