% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_postspawn.R
\name{make_postspawn}
\alias{make_postspawn}
\title{Post-spawning survival}
\usage{
make_postspawn(iteroparity, nM)
}
\arguments{
\item{habitat}{A built-in \code{data.frame} containing habitat 
estimates (km2) for the scenario and region selected.}
}
\description{
Function used to estimate post-spawn survival 
from proportion of repeat spawners by latitude (Leggett and
Cascardden 1978, Bailey and Zydlewski 2013) and natural mortality
by life-history region (ASMFC 2020)
}
