% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_s_spawn.R
\name{make_s_spawn}
\alias{make_s_spawn}
\title{Make total spawning migration mortality}
\usage{
make_s_spawn(nM, s_postspawn)
}
\arguments{
\item{nM}{Instantaneous natural mortality rate. 
A numeric vector of length 1.}

\item{s_postspawn}{Post spawning mortality calculated from the
output of make_iteroparity and nM.}
}
\value{
A vector of length one containing pre-spawn survival.
}
\description{
Function used to derive total in-river adult survival during
the spawning migration from annual mortality and post-spawn survival.
}
