% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spawnrecruit.R
\name{make_spawnrecruit}
\alias{make_spawnrecruit}
\title{Make age-structured probability of recruit to first spawn.}
\usage{
make_spawnrecruit(river, sex = c("male", "female"))
}
\arguments{
\item{river}{River for which spawn recruit probabilities are requested.}

\item{sex}{Sex of fish. If not specified, then mean of
male and female spawn recruit probabilities are returned.}
}
\value{
A numeric vector of \code{length = length(max_age)} depending
on maximum age in the selected river (by life-history region).
}
\description{
The purpose of this function is to make age-specific
recruit to spawn probabilities for rivers listed in \code{\link{get_rivers}}
using built-in data sets for region-specific \code{\link{max_ages}} and
\code{\link{maturity}}.
}
\examples{
make_spawnrecruit(river = "Susquehanna")

}
\references{
Atlantic States Marine Fisheries Commission
}
