% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spawnrecruit_eel.R
\name{make_spawnrecruit_eel}
\alias{make_spawnrecruit_eel}
\title{Make spawnrecruit vectors for American eel}
\usage{
make_spawnrecruit_eel(
  river,
  sex = c("male", "female"),
  species = "EEL",
  custom_habitat = NULL
)
}
\arguments{
\item{river}{River for which maximum age is needed.}

\item{sex}{A character indicating \code{"male"}, \code{"female"},
or \code{"Pooled"} sex for fish.}

\item{custom_habitat}{A dataframe containing columns corresponding to the
those in the output from \code{\link{custom_habitat_template}}. The default,
\code{NULL} uses the default habitat data set for a given combination of
\code{species} and \code{river}.}
}
\value{
A numeric vector containing a single realization for proportion
of mature fish at each age, from age 1 to maximum age.
}
\description{
Simulate proportion of population that is mature
spawners from sex-specific maturity schedules for river herring
in \code{\link{maki_pars}}.
}
\details{
The primary use of this function is to simulate proportion of
mature spawners at each age in a population of American eel based on
region-specific probabilities of maturation at each age. Estuarine
maturity schedule was estimated from the general stock assessment model 
by ASMFC (2012) and inland maturity schedule was derived from data collected
from the Shenandoah River (Sheila Eyler, U.S. Fish & Wildlife Service, 
unpublished data.)
}
\examples{
make_spawnrecruit_eel(river = "Hudson", species = "EEL", sex = "female")

}
\references{
Atlantic States Marine Fisheries Commission. 2023. American Eel
benchmark stock assessment and peer-review report. ASMFC, Arlington, VA.
URL: https://asmfc.org/wp-content/uploads/2024/11/AmEelBenchmarkStockAssessment_PeerReviewReport_Aug2023.pdf
}
