% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spawnrecruit_rh.R
\name{make_spawnrecruit_rh}
\alias{make_spawnrecruit_rh}
\title{Make spawnrecruit vectors for river herring}
\usage{
make_spawnrecruit_rh(
  river,
  sex = c("male", "female"),
  species = c("ALE", "BBH"),
  custom_habitat = NULL
)
}
\arguments{
\item{river}{River for which maximum age is needed.}

\item{sex}{A character indicating \code{"male"}, \code{"female"},
or \code{"Pooled"} sex for fish.}

\item{species}{A character indicating species of river herring, either
\code{"ALE"} for alewife or \code{"BBH"} for blueback herring.}

\item{custom_habitat}{A dataframe containing columns corresponding to the
those in the output from \code{\link{custom_habitat_template}}. The default,
\code{NULL} uses the default habitat data set for a given combination of
\code{species} and \code{river}.}
}
\value{
A numeric vector containing a single realization for proportion
of mature fish at each age, from age 1 to maximum age.
}
\description{
Simulate proportion of population that is mature
spawners from sex-specific maturity schedules for river herring
in \code{\link{maki_pars}}.
}
\details{
The primary use of this function is to simulate proportion of
mature spawners at each age in a population of river-herring based on 
region-specific probabilities of maturation at each age estimated
from the Maki (YEAR) method in ASMFC (2024).
}
\examples{
make_spawnrecruit_rh(river = "Hudson", species = "BBH", sex = "female")

}
\references{
Atlantic States Marine Fisheries Commission. 2024. River herring
benchmark stock assessment and peer-review report. ASMFC, Arlington, VA. 
URL: https://asmfc.org/uploads/file/66f59e40RiverHerringAssessment_PeerReviewReport_2024.pdf
}
