% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maki_pars.R
\docType{data}
\name{maki_pars}
\alias{maki_pars}
\title{Sex-specific maturity schedules for River Herring}
\format{
A data frame with 216 observations of 6 variables:
\describe{

    \item{\code{Species}}{Species of fish}
    \item{\code{Region}}{Genetic reporting group and general geographic region}
    \item{\code{Sex}}{Fish sex}
    \item{\code{Age}}{Fish age (years)}
    \item{\code{pMature}}{Probability of maturation at age}
    \item{\code{pMat.se}}{Standard error of estimated pMature from ASMFC (2024)}
}
}
\source{
Atlantic States Marine Fisheries Commission
}
\usage{
maki_pars
}
\description{
A dataset containing sex-specific probabilities 
of maturation at age for each life-history region 
(SAT = Southern Atlantic, MAT = Mid Atlantic, 
SNE = Southern New England, MNE = Mid New England,
NNE = Northern New England, CAN-NNE = Canada and Northern New England)
}
\references{
Atlantic States Marine Fisheries Commission. 2024. River herring
benchmark stock assessment and peer-review report. ASMFC, Arlington, VA. 
URL: https://asmfc.org/uploads/file/66f59e40RiverHerringAssessment_PeerReviewReport_2024.pdf
}
\keyword{datasets}
