% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturity.R
\docType{data}
\name{maturity}
\alias{maturity}
\title{Sex-specific maturity schedules for American shad}
\format{
A data frame with 6 observations of 15 variables:
\describe{
    \item{\code{region}}{Life-history region}
    \item{\code{sex}}{Fish sex}
    \item{\code{age1}}{Probability of maturing at age 1}
    \item{\code{age2}}{Probability of maturing at age 2}
    \item{\code{age3}}{Probability of maturing at age 3}
    \item{\code{age4}}{Probability of maturing at age 4}
    \item{\code{age5}}{Probability of maturing at age 5}
    \item{\code{age6}}{Probability of maturing at age 6}
    \item{\code{age7}}{Probability of maturing at age 7}
    \item{\code{age8}}{Probability of maturing at age 8}
    \item{\code{age9}}{Probability of maturing at age 9}
    \item{\code{age10}}{Probability of maturing at age 10}
    \item{\code{age11}}{Probability of maturing at age 11}
    \item{\code{age12}}{Probability of maturing at age 12}
    \item{\code{age13}}{Probability of maturing at age 13}
}
}
\source{
Atlantic States Marine Fisheries Commission
}
\usage{
maturity
}
\description{
A dataset containing sex-specific probabilities
of maturation at age for each life-history region
(NI = Northern iteroparous, SI = Southern iteroparous,
SP = Semelparous)
}
\references{
Atlantic States Marine Fisheries Commission (ASMFC). 2020.
American shad benchmark stock assessment and peer-review report. ASMFC,
Arlington, VA.
}
\keyword{datasets}
