% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturity.R
\docType{data}
\name{maturity}
\alias{maturity}
\title{Sex-specific maturity schedules for American shad}
\format{
A data frame with 6 observations of 15 variables:
\describe{

\code{region} Life-history region

\code{sex} Fish sex

\code{age1, age2, age3, ...} Probability of maturing at age
}
}
\source{
Atlantic States Marine Fisheries Commission
}
\usage{
maturity
}
\description{
A dataset containing sex-specific probabilities 
of maturation at age for each life-history region 
(NI = Northern iteroparous, SI = Southern iteroparous, 
SP = Semelparous)
}
\references{
Atlantic States Marine Fisheries Commission (ASMFC). 2020. 
American shad benchmark stock assessment and peer-review report. ASMFC, 
Arlington, VA.
}
\keyword{datasets}
