% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_ages.R
\docType{data}
\name{max_ages}
\alias{max_ages}
\title{Sex-specific maximum ages of American shad}
\format{A data frame with 6 observations of 3 variables:
\describe{

\code{region}{ Life-history region}

\code{sex}{ Fish sex (gender)}

\code{maxage}{ Maximum observed age}

}}
\source{
Atlantic States Marine Fisheries Commission
}
\usage{
max_ages
}
\description{
A dataset containing sex-specific maximum
ages for each life-history region (NI = Northern iteroparous, 
SI = Southern iteroparous, SP = Semelparous)
}
\keyword{datasets}
