% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\docType{data}
\name{mortality}
\alias{mortality}
\title{Sex-specific instaneous mortality of American shad}
\format{
A data frame with 6 observations of 3 variables:
\describe{
    \item{\code{region}}{Life-history region}
    \item{\code{sex}}{Fish sex}
    \item{\code{M}}{Instantaneous natural mortality}
}
}
\source{
Atlantic States Marine Fisheries Commission
}
\usage{
mortality
}
\description{
A dataset containing sex-specific instantaneous
mortality estimates for each life-history region 
(NI = Northern iteroparous, SI = Southern iteroparous, 
SP = Semelparous).
}
\references{
Atlantic States Marine Fisheries Commission (ASMFC). 2020. 
American shad benchmark stock assessment and peer-review report. ASMFC, 
Arlington, VA.
}
\keyword{datasets}
