% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_rh.R
\docType{data}
\name{mortality_rh}
\alias{mortality_rh}
\title{Sex-specific instantaneous natural mortality of river herring}
\format{
A data frame with 300 observations of 5 variables:
\describe{
    \item{\code{region}}{Genetic reporting group region}
    \item{\code{sex}}{Sex of fish}
    \item{\code{age}}{Age in years}
    \item{\code{M}}{Instantaneous natural mortality}
    \item{\code{species}}{Species}
}
}
\source{
Atlantic States Marine Fisheries Commission
}
\usage{
mortality_rh
}
\description{
A dataset containing sex- and age-specific instantaneous
natural mortality estimates for each species and genetic reporting region
}
\references{
Atlantic States Marine Fisheries Commission. 2024. River herring
benchmark stock assessment and peer-review report. ASMFC, Arlington, VA. 
URL: https://asmfc.org/uploads/file/66f59e40RiverHerringAssessment_PeerReviewReport_2024.pdf
}
\keyword{datasets}
