% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_pop.R
\name{project_pop}
\alias{project_pop}
\title{Project population}
\usage{
project_pop(x, age0, nM, fM)
}
\arguments{
\item{x}{An numeric vector containing age-specific 
abundance at end of year t.}

\item{age0}{A vector of length one containing number of new recruits 
at end of year t.}

\item{nM}{Instantaneous natural mortality.}

\item{fM}{Instantaneous fishing mortality.}
}
\value{
A numeric vector of age-structured abundance in time t + 1

#@example inst/examples/makefec_ex.R
}
\description{
Function used to project population forward 
one time step using matrix multiplication. Based on 
modification of the \code{project.leslie} in \code{demogR}
}
