% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_pop.R
\name{project_pop}
\alias{project_pop}
\title{Project population}
\usage{
project_pop(x, age0, nM, fM, max_age, species = c("AMS", "ALE", "BBH"))
}
\arguments{
\item{x}{A numeric vector containing age-specific 
abundance at end of year t.}

\item{age0}{A numeric vector of length one containing number of new recruits 
at end of year t.}

\item{nM}{Numeric vector of instantaneous natural mortality.}

\item{fM}{Numeric vector of instantaneous fishing mortality.}

\item{max_age}{Numeric indicating maximum age of spawning fish.}

\item{species}{Species for which population dynamics will be simulated.
Choices include American shad (\code{"AMS"}), alewife (\code{"ALE"}), and
blueback herring (\code{"BBH"}).}
}
\value{
A numeric vector of age-structured abundance in time t + 1.
}
\description{
Function used to project population forward 
one time step using vector multiplication.
}
\examples{
# Example usage
\dontrun{
  
 # Simulate a population
 n_init = 1e5  
 max_age = 13
 nM = .505
 fM = 0
 pop = make_pop(max_age=max_age, nM=nM, fM=fM, n_init=n_init)
 
 # Project one time-step without reproduction
 new_pop = project_pop(x=pop, age0=1e5, nM=.505, fM=0, max_age)
 
 # Difference should be about zero 
 # because age0 - n_init
 new_pop - pop
  
}
}
