% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_juvenile_s.R
\name{sim_juvenile_s}
\alias{sim_juvenile_s}
\title{Juvenile (hatch-to-outmigrant) survival}
\usage{
sim_juvenile_s(crecco_1983)
}
\arguments{
\item{crecco_1983}{built-in dataset containing daily mortality 
rates, daily survival, and cumulative survival during varying
post-hatch periods. See \code{\link{crecco_1983}}.}
}
\description{
Function used to simulate juvenile (hatch-to-outmigrant) 
survival from daily rates reported by Crecco et al. (1983) from
the Connecticut River, USA during 4 consecutive years.
}
\references{
Crecco, V., T. Savoy, and L. Gunn. 1983. Daily mortality rates 
of larval and juvenile American shad (*Alosa sapidissima*) in the Connecticut 
River with changes in year-class strength. Canadian Journal of 
Fisheries and Aquatic Sciences 40:1719–1728.
}
