% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_juvenile_s.R
\name{sim_juvenile_s}
\alias{sim_juvenile_s}
\title{Juvenile (hatch-to-outmigrant) survival}
\usage{
sim_juvenile_s(species = c("AMS", "ALE", "BBH"))
}
\arguments{
\item{species}{Species used for simulation ("AMS", "ALE", or "BBH")}
}
\description{
Function used to simulate juvenile (hatch-to-outmigrant) 
survival from daily rates of A
}
\examples{
sim_juvenile_s()

}
\references{
Crecco, V., T. Savoy, and L. Gunn. 1983. Daily mortality rates 
of larval and juvenile American shad (*Alosa sapidissima*) in the Connecticut 
River with changes in year-class strength. Canadian Journal of 
Fisheries and Aquatic Sciences 40:1719–1728.

Overton et al. (2012)

Stich et al. (in press)

Hook et al. (2007)
}
