% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_pop}
\alias{sim_pop}
\title{Simulate population dynamics through time.}
\usage{
sim_pop(nyears, river, max_age = NULL, nM = NULL, fM, n_init,
  spawnRecruit = NULL, eggs = NULL, sr, s_prespawn,
  s_juvenile = NULL, upstream, downstream, downstream_j,
  output_years = "last", age_structured_output = FALSE)
}
\arguments{
\item{nyears}{Number of years for simulation(s).}

\item{river}{River basin. Available rivers implemented in package 
can be viewed by calling \code{\link{get_rivers}}. Alternatively,
the user can specify \code{rivers = sample(get_rivers, 1)} 
to randomly sample river within larger simulation studies. Information
about each river can be found in \code{\link{inventory}}, 
\code{\link{shad_rivers}}, and \code{\link{habitat}} datasets.}

\item{max_age}{Maximum age of fish in population. If \code{NULL}
(default), then based on the maximum age of females for the
corresponding region in the \code{\link{max_ages}} dataset.}

\item{nM}{Instantaneous natural mortality. If \code{NULL}
(default), then based the average of males and females for the
corresponding region in the \code{\link{mortality}} dataset.}

\item{fM}{Instantaneous fishing mortality. The default value is zero.}

\item{n_init}{Initial population seed (number of Age-1 individuals) 
used to simulate the starting population.}

\item{spawnRecruit}{Probability of recruitment to spawn at age. Numeric 
vector with length \code{max_age}}

\item{eggs}{Number of eggs per female. Can be a vector of length 1
if eggs per female is age invariant, or can be vector of length 
\code{max_age} if age-specific. If \code{NULL} (default) then 
estimated based on regression relationships in \code{\link{make_eggs}}.}

\item{sr}{Sex ratio (expressed as percent female or P(female)).}

\item{s_prespawn}{Pre-spawn survival for spawners.}

\item{s_juvenile}{Survival from hatch to outmigrant. If NULL
(default) then simulated from a (log) normal distribution using
mean and sd of \code{Sc} through \code{70 d} for 1979-1982 from
Crecco et al. (1983).}

\item{upstream}{Numeric of length 1 representing proportional 
upstream passage through dams.}

\item{downstream}{Numeric of length 1 indicating proportional 
downstream survival through dams.}

\item{downstream_j}{Numeric of length 1 indicating proportional
downstream survival through dams for juveniles.}

\item{output_years}{Level of detail provided in output. The default
value of '\code{last}' returns the final year of simulation.
Any value other than the default '\code{last}' will return
data for all years of simulation. This is useful for testing.}

\item{age_structured_output}{Should population and spawner abundance
in the output data.frame be age-structured? If \code{FALSE} (default),
then \code{pop} (non-spawning population) and \code{spawners} (spawning
population) are summed across all age classes for each year of simulation.
If \code{TRUE} then \code{pop} and \code{spawners} are returned for
each age class. For the sake of managing outputs, abundances for
\code{pop} and \code{spawners} are reported for all age classes 1-13 
regardless of \code{max_age}, but all abundances for ages greater 
than \code{max_age} are zero.}
}
\value{
A data.frame containing simulation inputs (arguments
to \code{sim_pop}) and output (number of spawners) by year.
}
\description{
Use functions from \code{\link{anadrofish}} to simulate
population change through time.
}
