% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_pop}
\alias{sim_pop}
\title{Simulate population dynamics through time.}
\usage{
sim_pop(nyears, river, max_age = NULL, nM = NULL, fM, n_init,
  spawnRecruit = NULL, eggs = NULL, sr, s_prespawn, s_hatch, upstream,
  downstream, downstream_j, output = "last")
}
\arguments{
\item{nyears}{Number of years for simulation(s).}

\item{river}{River basin. Available rivers implemented in package 
can be viewed by calling \code{\link{get_rivers}}. Alternatively,
the user can specify \code{rivers = sample(get_rivers, 1)} 
to randomly sample river within larger simulation studies. Information
about each river can be found in \code{\link{inventory}} and 
\code{\link{shad_rivers}} datasets.}

\item{max_age}{Maximum age of fish in population. If \code{NULL}
(default), then based on the maximum age of females for the
corresponding region in the \code{\link{max_ages}} dataset.}

\item{nM}{Instantaneous natural mortality. If \code{NULL}
(default), then based the average of males and females for the
corresponding region in the \code{\link{mortality}} dataset.}

\item{fM}{Instantaneous fishing mortality. The default value is zero.}

\item{n_init}{Initial population seed (number of Age-1 individuals) 
used to simulate the starting population.}

\item{spawnRecruit}{Probability of recruitment to spawn at age. Numeric 
vector with length \code{max_age}}

\item{eggs}{Number of eggs per female. Can be a vector of length 1
if eggs per female is age invariant, or can be vector of length 
\code{max_age} if age-specific. If \code{NULL} (default) then 
estimated based on regression relationships in \code{\link{make_eggs}}.}

\item{sr}{Sex ratio (expressed as percent female or P(female)).}

\item{s_prespawn}{Pre-spawn survival for spawners.}

\item{s_hatch}{Survival from hatch to outmigrant.}

\item{upstream}{Numeric of length 1 representing proportional 
upstream passage through dams.}

\item{downstream}{Numeric of length 1 indicating proportional 
downstream survival through dams.}

\item{downstream_j}{Numeric of length 1 indicating proportional
downstream survival through dams for juveniles.}

\item{output}{Level of detail provided in output. The default
value of '\code{last}' returns the final year of simulation.
Any value other than the default '\code{last}' will return
data for all years of simulation. This is useful for testing.}
}
\value{
A data.frame containing simulation inputs (arguments
to \code{sim_pop}) and output (number of spawners) by year.
}
\description{
Use functions from \code{\link{anadrofish}} to simulate
population change through time.
}
