% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_pop}
\alias{sim_pop}
\title{Simulate population dynamics through time}
\usage{
sim_pop(nyears = 100, river = "Penobscot", max_age = 9, nM = 0.38,
  fM = 0, n_init = 1e+05, spawnRecruit = c(0, 0, 0, 0.01, 0.33, 0.84,
  0.97, 0.99, 1), eggs = c(0, 0, 0, 20654, 34674, 58210, 79433, 88480,
  97724), sr = 0.5, s_prespawn = 0.9, s_hatch = 0.1,
  params = c("river", "max_age", "nM", "fM", "n_init", "spawnRecruit",
  "eggs", "sr", "s_hatch", "S_prespawn", "spawners", "pop"))
}
\arguments{
\item{nyears}{Number of years.}

\item{river}{River basin. Currently defined as HUC 10 watershed.}

\item{max_age}{Maximum age of fish in population.}

\item{nM}{Instantaneous natural mortality estimate.}

\item{fM}{Instantaneous fishing mortality estimate.}

\item{n_init}{Initial population seed (number of Age-1 individuals) 
used to seed the population.}

\item{spawnRecruit}{Probability of recruitment to spawn at age. Numeric 
vector with length \code{max_age}}

\item{eggs}{Number of eggs per female. Can be a vector of length 1
if eggs per female is age invariant, or can be vector of length 
\code{max_age} if age-specific.}

\item{sr}{Sex ratio (expressed as percent female or P(female)).}

\item{s_prespawn}{Pre-spawn survival for spawners.}

\item{s_hatch}{Survival from hatch to outmigrant}
}
\value{
A data.frame containing simulation inputs (arguments) 
and outputs (pop, spawners) by year.
}
\description{
Use functions from anadrofish to simulate
population change through time.
}
\examples{
# Example usage
\dontrun{
  
  res <- sim_pop(
    nyears = 50,
    river = 'Penobscot',
    max_age = 9,
    nM = 0.38,
    fM = 0.00,
    n_init = 1e5,
    spawnRecruit = c(0, 0, 0, 0.01, 0.33, 0.84, 0.97, 0.99, 1.00), 
    eggs = c(0, 0, 0, 20654, 34674, 58210, 79433, 88480, 97724),
    sr = 0.50,
    s_prespawn = 0.90,  
    s_hatch = 0.10)
  
  plot(x=res$out_year, y = apply(res[ , c(27:35)], 1, sum) , typ='l')
  
}
}
