% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbgf_bbh.R
\docType{data}
\name{vbgf_bbh}
\alias{vbgf_bbh}
\title{Sex-specific von Bertalanffy growth parameters for 
blueback herring}
\format{
A data frame with 816,000 observations of 5 variables:
\describe{
    \item{\code{Region}}{Genetic reporting group}
    \item{\code{Sex}}{Sex of fish}
    \item{\code{Linf}}{Asymptotic length of fish}
    \item{\code{K}}{Brody growth coefficient}
    \item{\code{t0}}{Size at age zero}
}
}
\source{
Atlantic States Marine Fisheries Commission
}
\usage{
vbgf_bbh
}
\description{
A dataset containing sex-specific estimates of von Bertalanffy
growth parameters for blueback herring from the 2024 ASMFC river herring 
benchmark stock assessment (ASMFC 2024).
}
\references{
Atlantic States Marine Fisheries Commission. 2024. River herring
benchmark stock assessment and peer-review report. ASMFC, Arlington, VA. 
URL: https://asmfc.org/uploads/file/66f59e40RiverHerringAssessment_PeerReviewReport_2024.pdf
}
\keyword{datasets}
