% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_output}
\alias{write_output}
\title{Write simulation results}
\usage{
write_output()
}
\value{
A list of results.
}
\description{
Internal function used to collect and
return all model inputs and relevant model
outputs (e.g., population size) from all years t.

Not intended to be called directly.
}
