% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_output}
\alias{write_output}
\title{Write simulation results}
\usage{
write_output(.sim_pop)
}
\arguments{
\item{.sim_pop}{A hidden environment in the calling frame 
of \code{\link{sim_pop}}. Arbitrarily any named object with
names matching those used in the function.}
}
\value{
A list of results.
}
\description{
Internal function used to collect and
return all model inputs and relevant model
outputs (e.g., population size) from all years t.

For internal use in \code{\link{sim_pop}}. Not intended 
to be called directly.
}
\keyword{Internal}
